% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_community.R
\name{make_community}
\alias{make_community}
\title{Simulate microbial community on a 2D space}
\usage{
make_community(parameters)
}
\arguments{
\item{parameters}{contains the list of parameters to build the community. See \code{Details}.}
}
\value{
a marked point process with type (isolated, microcolony, filamentous) and parent cluster id (column phylum)
}
\description{
The function \code{make_community} distributes points on a 2D space to simulate the distribution of microbial cells on pore surfaces.
}
\details{
The \code{parameters} list should contain:
\describe{
\item{\code{dimxy}}{the dimension of the square window}
\item{\code{prop_types}}{the proportions of microcolonies, filamentous and isolated cells}
\item{\code{cell_diameter}}{the diameters of cells (identical for all groups)}
\item{\code{microcolonies_size_range}}{the range for cell numbers in microcolonies}
\item{\code{microcolonies_diameter}}{the range for microcolony diameter}
\item{\code{microcolonies_distribution}}{the type of spatial distribution for microcolonies. Choose between \emph{NeymanScott}, \emph{Matern} or \emph{Thomas}}
\item{\code{filamentous_size_range}}{the the range for cell numbers in filamentous colonies}
}
}
\examples{
parms = list(dimxy = 100,
             prop_types = c(40,50,10),
             cell_density = 0.0002,
             cell_diameter = 1,
             microcolonies_size_range = c(8,10),
             microcolonies_diameter = 10,
             microcolonies_distribution = "Matern",
             filamentous_size_range = c(8,10)
             )
make_community(parms)
}
