% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_species.R
\name{populate_species}
\alias{populate_species}
\title{Add phylum marks to microbila community made with \code{make_community()}}
\usage{
populate_species(X, props)
}
\arguments{
\item{X}{a simulated community produced by \code{make_community()}}

\item{props}{a named vector of proportions for the different phyla. Must add to 1.}
}
\description{
This function adds a mark to each point: the group (phylum) to which it belong.
}
\details{
Depending on the total number of cells and the proportion in each group, we may not obtain round numbers. If the total number of cells does not allow to have integer groups, the function samples the groups with given proportion. This means that realized proportions can deviate from the given proportions. This is especially the case when cell density is low.
In the case of colonies, cells within a colonie are assigned to the same phylum
}
\examples{
parms = list(dimxy = 100,
             prop_types = c(40,50,10),
             cell_density = 0.0002,
             cell_diameter = 1,
             microcolonies_size_range = c(8,10),
             microcolonies_diameter = 10,
             microcolonies_distribution = "Matern",
             filamentous_size_range = c(8,10)
)
populate_species(make_community(parms) , c(phylum1 = 0.20,
                                           phylum2 = 0.40, 
                                           phylum3 = 0.10,
                                           phylum4 = 0.25, 
                                           phylum5 = 0.05)
                                           )
}
